requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
scipy==1.11.4
pydantic==2.5.2
python-dotenv==1.0.0
aiosqlite==0.19.0
asyncio==3.4.3
python-multipart==0.0.6
httpx==0.25.2
joblib==1.3.2
colorlog==6.8.0
pytz==2023.3.post1